feat: Session 7 — callFunction, ffe.requestShutdown, jitter fix (PreviousTransform)
ScriptEngine::callFunction(funcName, entityId, dt) — per-frame Lua function
invocation via lua_getglobal + lua_pcall with no per-call recompilation.
Resolves FRICTION-1 HIGH from game-dev-tester (doString compile overhead).

ffe.requestShutdown() Lua binding — scripts signal clean engine shutdown via
ShutdownSignal in ECS registry context. Null-guard added after security-auditor
review to prevent UB when no World is registered.

Jitter root cause fixed: interpolation alpha was discarded in Application::render().
Implemented Option A from design-note-interpolation.md:
- PreviousTransform component (mirrors Transform fields)
- copyTransformSystem (priority 5, runs before gameplay each tick)
- renderPrepareSystem moved outside registered system list; called from
  Application::render(alpha) with correct interpolation alpha
- Two-pass lerp: entities with PreviousTransform get smoothed positions;
  entities without use raw current position

All examples (hello_sprites, interactive_demo, lua_demo) updated with
PreviousTransform on moving entities.

177/177 tests pass on Clang-18 and GCC-13. Zero warnings.

+++ engine/renderer/opengl/rhi_opengl.cpp
@@ -359,7 +359,8 @@ RhiResult updateBuffer(const BufferHandle handle, const void* data, const u32 si
     }
 
     const GlBuffer& buf = s_buffers[handle.id];
-    if (offset + sizeBytes > buf.sizeBytes) {
+    // M-7: cast to u64 before addition to prevent u32 overflow
+    if (static_cast<u64>(offset) + static_cast<u64>(sizeBytes) > static_cast<u64>(buf.sizeBytes)) {
         FFE_LOG_ERROR("Renderer", "Buffer update exceeds buffer size");
         return RhiResult::ERROR_INVALID_HANDLE;
     }
